+++ tests/test_compute_everything.py
@@ -8,6 +8,7 @@ import pytest
 
 from desc.coils import (
     FourierPlanarCoil,
+    FourierPlanarFiniteBuildCoil,
     FourierRZCoil,
     FourierXYCoil,
     FourierXYZCoil,
@@ -188,6 +189,14 @@ def test_compute_everything():
         ),
         "desc.coils.SplineXYZCoil": SplineXYZCoil(
             current=5, X=[5, 10, 2, 5], Y=[1, 2, 3, 1], Z=[-4, -5, -6, -4]
+        ),
+        "desc.coils.FourierPlanarFiniteBuildCoil": FourierPlanarFiniteBuildCoil(
+            current=5,
+            center=[10, 1, 3],
+            normal=[1, 2, 3],
+            r_n=[1, 2, 3],
+            modes=[0, 1, 2],
+            cross_section_dims=[0.1, 0.2],
         ),
     }
     assert things.keys() == data_index.keys(), (