+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Optionally can also contract the profiles of the original ``Equilibrium`` so tha
 - Adds ``source_grid`` argument to ``desc.magnetic_fields._MagneticField.save_mgrid function`` to allow user to control the discretization of the magnetic field object being used to construct the ``mgrid`` output.
 - Removes default objective and constraints for ``desc.equilibrium.Equilibrium.optimize``, so now user is required to pass in the constraints and objective when using this method.
 - Adds ``PlasmaCoilSetDistanceBound`` objective to allow a simultaneous constraint on minimum and maximum distance between the plasma and coils.
+- Adds support for self-field computation with finite build coils in the ``AbstractFiniteBuildCoil`` class, and implements ``FourierPlanarFiniteBuildCoil``. Adds ``CoilSetMaxB``
+objective to constrain maximum field on coil in optimization.
 
 Performance Improvements